requirements.txt
beautifulsoup4==4.12.3
langchain_community==0.3.0
langchain_core==0.3.3
langchain_openai==0.2.0
langgraph==0.2.22
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
termcolor==2.4.0
